fix portrait mode
