+++ src/app/exhibition-info/osaka/creators-market/styles/map.css
@@ -272,7 +272,7 @@ button.booth {
   z-index: 10000;
 }
 .booth .booth-label {
-  display: inline-block;
+  display: inline-flex;
   transform: rotate(0deg);
   transform-origin: center;
 }
@@ -285,24 +285,10 @@ button.booth {
 }
 
 @media (max-width:480px){
-  .cm-map-wrapper .entrance-label{
-    transform:rotate(-90deg);
-  }
   .cm-grid{
     width:100%;
     height:100%;
     grid-auto-rows:1fr;
-  }
-  .cm-grid .booth-empty,
-  .cm-grid .exit-label-cell{
-    transform:rotate(-90deg);
-    transform-origin:center;
-  }
-  .cm-grid .booth .booth-label{
-    display:inline-block;
-  }
-  .cm-grid .booth .booth-label span{
-    display:block;
   }
 }
 
@@ -315,6 +301,7 @@ button.booth {
     width:100%;
     max-width:100%;
     box-sizing:border-box;
+    margin:0;
   }
   .cm-map-wrapper::before,
   .cm-map-wrapper::after{
@@ -328,7 +315,7 @@ button.booth {
     content:none;
   }
   .cm-grid .booth .booth-label{
-    transform:rotate(-90deg);
-    transform-origin:center;
+    flex-direction:column;
+    align-items:center;
   }
 }
