+++ src/styles/portraitMapLabelRotation.css
@@ -1,0 +1,7 @@
+// 가로모드 모바일 레거시코드
+@media (max-width:480px) and (orientation:portrait){
+  .cm-grid .booth .booth-label{
+    transform:rotate(-90deg);
+    transform-origin:center;
+  }
+}
